remove unused ingestion and metrics
