+++ engine/testutil/mock/nodes.go
@@ -22,7 +22,7 @@ import (
 	"github.com/onflow/flow-go/engine/consensus/matching"
 	"github.com/onflow/flow-go/engine/consensus/sealing"
 	"github.com/onflow/flow-go/engine/execution/computation"
-	"github.com/onflow/flow-go/engine/execution/ingestion"
+	executionIngest "github.com/onflow/flow-go/engine/execution/ingestion"
 	executionprovider "github.com/onflow/flow-go/engine/execution/provider"
 	"github.com/onflow/flow-go/engine/execution/state"
 	"github.com/onflow/flow-go/engine/verification/assigner"
@@ -192,7 +192,7 @@ func (cn ConsensusNode) Done() {
 type ExecutionNode struct {
 	GenericNode
 	FollowerState       protocol.FollowerState
-	IngestionEngine     *ingestion.Engine
+	IngestionEngine     *executionIngest.Core
 	ExecutionEngine     *computation.Manager
 	RequestEngine       *requester.Engine
 	ReceiptsEngine      *executionprovider.Engine
@@ -217,6 +217,7 @@ func (en ExecutionNode) Ready(ctx context.Context) {
 	// new interface.
 	irctx, _ := irrecoverable.WithSignaler(ctx)
 	en.ReceiptsEngine.Start(irctx)
+	en.IngestionEngine.Start(irctx)
 	en.FollowerCore.Start(irctx)
 	en.FollowerEngine.Start(irctx)
 	en.SyncEngine.Start(irctx)
@@ -238,7 +239,6 @@ func (en ExecutionNode) Done(cancelFunc context.CancelFunc) {
 
 	// to stop all (deprecated) ready-done-aware
 	<-util.AllDone(
-		en.IngestionEngine,
 		en.IngestionEngine,
 		en.ReceiptsEngine,
 		en.Ledger,
