+++ module/metrics/noop.go
@@ -193,7 +193,6 @@ func (nc *NoopCollector) ReadValuesSize(byte uint64)                            
 func (nc *NoopCollector) ReadDuration(duration time.Duration)                                   {}
 func (nc *NoopCollector) ReadDurationPerItem(duration time.Duration)                            {}
 func (nc *NoopCollector) ExecutionCollectionRequestSent()                                       {}
-func (nc *NoopCollector) ExecutionCollectionRequestRetried()                                    {}
 func (nc *NoopCollector) RuntimeTransactionParsed(dur time.Duration)                            {}
 func (nc *NoopCollector) RuntimeTransactionChecked(dur time.Duration)                           {}
 func (nc *NoopCollector) RuntimeTransactionInterpreted(dur time.Duration)                       {}
